+++ CHANGES.md
@@ -1,5 +1,12 @@
 #### [unreleased]
 
+#### 5.2.0 / 2018-07-28
+* write standard and minified CSS to files
+* load minified CSS when WP_DEBUG is false
+* improve minify code
+* update CSS selectors for list view
+* refactor CSS strings for inline format
+
 #### 5.1.0 / 2018-07-21
 * update for new tooltip title as `h3`
 * added some functions for adding deprecated CSS